A few minor changes
placeholder mass resolution for 2.3 GeV is 63% percent of the formula
for mass resolution that is in Sho's dissertation.  This will be changed
to something more realistic once I get the A' MC from Bradley

Renamed the plots in the output file, so that it's easier to tell which
plot is which.